add redis config

+++ src/main/java/com/example/demo/dao/ClientDao.java
@@ -8,6 +8,6 @@ public interface ClientDao {
     public Client addClient(Client client);
     public Client updateClient(Client client);
     public void removeClient(Client client);
-    public Client getClientById(int id);
+    public Client getClientById(long id);
     public List<Client> listClients();
 }
